Move code to Hash.swift
