+++ ios/ReactNativeWalletCore.swift
@@ -22,47 +22,4 @@ class ReactNativeWalletCore: NSObject {
         }
         resolve(wallet.mnemonic)
     }
-    
-    @objc(sha3_256:withResolver:withRejecter:)
-    func sha3_256(string: String, resolve:RCTPromiseResolveBlock, reject:RCTPromiseRejectBlock) -> Void {
-        guard let data = stringToData(string: string) else {
-            reject(Self.errorDomain, "invalid hex string", nil)
-            return
-        }
-        resolve(Hash.sha3_256(data: data).hexString)
-    }
-    
-    @objc(sha256:withResolver:withRejecter:)
-    func sha256(string: String, resolve:RCTPromiseResolveBlock, reject:RCTPromiseRejectBlock) -> Void {
-        guard let data = stringToData(string: string) else {
-            reject(Self.errorDomain, "invalid hex string", nil)
-            return
-        }
-        resolve(Hash.sha256(data: data).hexString)
-    }
-    
-    @objc(sha512:withResolver:withRejecter:)
-    func sha512(string: String, resolve:RCTPromiseResolveBlock, reject:RCTPromiseRejectBlock) -> Void {
-        guard let data = stringToData(string: string) else {
-            reject(Self.errorDomain, "invalid hex string", nil)
-            return
-        }
-        resolve(Hash.sha512(data: data).hexString)
-    }
-    
-    @objc(keccak256:withResolver:withRejecter:)
-    func keccak256(string: String, resolve:RCTPromiseResolveBlock, reject:RCTPromiseRejectBlock) -> Void {
-        guard let data = stringToData(string: string) else {
-            reject(Self.errorDomain, "invalid hex string", nil)
-            return
-        }
-        resolve(Hash.keccak256(data: data).hexString)
-    }
-    
-    func stringToData(string: String) -> Data? {
-        if string.hasPrefix("0x") {
-            return Data(hexString: string)
-        }
-        return Data(string.utf8)
-    }
 }
